skills/ocr-skill/pyproject.toml
[project]
name = "ocr-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pytesseract", "Pillow"]
